backend/requirements.txt
Flask
opencv-python
Pillow
numpy
python-dateutil
tensorflow
scikit-learn
easyocr
flask-cors
gunicorn
mtcnn